+++ src/components/Carousel.tsx
@@ -4,12 +4,13 @@ import 'swiper/css';
 import 'swiper/css/mousewheel';
 import 'swiper/css/scrollbar';
 import 'swiper/css/effect-coverflow';
-import { Mousewheel, Scrollbar, EffectCoverflow, FreeMode } from 'swiper/modules';
+import { Mousewheel, Scrollbar, EffectCoverflow } from 'swiper/modules';
 import SavedNews from './SavedNews';
 
-const Carousel: React.FC = () => {
+const Carousel = () => {
   const swiperRef = useRef(null);
 
+  // 페이지 전체 스크롤 방지
   useEffect(() => {
     const originalStyle = window.getComputedStyle(document.body).overflow;
     document.documentElement.style.overflow = 'hidden';
@@ -21,12 +22,12 @@ const Carousel: React.FC = () => {
     };
   }, []);
 
+  // 마우스 휠 이벤트
   useEffect(() => {
     const handleWheel = (e) => {
       if (swiperRef.current && swiperRef.current.swiper) {
-        const swiper = swiperRef.current.swiper;
-        swiper.mousewheel.enable();
-        swiper.mousewheel.handle(e);
+        swiperRef.current.swiper.mousewheel.enable();
+        swiperRef.current.swiper.mousewheel.handle(e);
       }
     };
 
@@ -37,23 +38,25 @@ const Carousel: React.FC = () => {
     };
   }, []);
 
-  const newsData = Array.from({ length: 20 }, (_, index) => ({ id: index + 1 }));
+  // 슬라이드 데이터 갯수
+  const newsData = Array.from({ length: 10 }, (_, index) => ({ id: index + 1 }));
 
-  const spaceBetweenInPx = 1.125 * 16; // 1.125rem을 픽셀 값으로 변환 (18px)
-
+  // 슬라이드 변경 시 슬라이드 크기 및 투명도 업데이트
   const updateSlideOpacityAndSize = (swiper) => {
     const slides = swiper.slides;
     slides.forEach((slide, index) => {
       slide.style.transition = 'opacity 0.5s ease, transform 0.5s ease';
       if (index === swiper.activeIndex) {
         slide.style.opacity = '1';
-        slide.style.transform = 'scale(1)'; // 활성 슬라이드 크기 증가
+        slide.style.transform = 'w-[65rem] h-[15rem]'; // 활성 슬라이드 크기 증가
       } else if (index === swiper.activeIndex - 1 || index === swiper.activeIndex + 1) {
-        slide.style.opacity = '0.8';
+        slide.style.opacity = '0.7';
         slide.style.transform = 'w-[60rem] h-[10rem]'; // 인접한 슬라이드 크기 증가
+        // slide.style.transform = 'scale(0.8)';
       } else if (index === swiper.activeIndex - 2 || index === swiper.activeIndex + 2) {
-        slide.style.opacity = '0.6';
-        slide.style.transform = 'w-[55rem] h-[5rem]'; // 그 다음 슬라이드 기본 크기
+        slide.style.opacity = '0.5';
+        slide.style.transform = 'w-[55rem] h-[5rem]'; // 인접한 슬라이드 크기 증가
+        // slide.style.transform = 'scale(0.5)'; // 그 다음 슬라이드 기본 크기
       } else {
         slide.style.opacity = '0';
         slide.style.transform = 'scale(0)'; // 그 외 슬라이드 크기 감소
@@ -61,26 +64,25 @@ const Carousel: React.FC = () => {
     });
   };
 
+  // 슬라이드 클릭시 뭘 할건지
+  const handleSlideClick = (index) => {
+    console.log(`Slide ${index} clicked`);
+  };
+
   return (
     <div className="flex justify-center items-center w-screen h-screen overflow-hidden">
-      <div className="relative w-full h-full" style={{ width: '1068px', height: '792px' }}>
+      <div className="relative w-full h-full" style={{ width: '66.75rem', height: '49.5rem' }}>
         <style>{`
           .swiper-scrollbar {
             background: transparent;
             position: absolute;
-
           }
           .swiper-scrollbar-drag {
             background: black;
           }
-          @media (max-width: 768px) {
-            .swiper-scrollbar {
-              right: 10px; /* 작은 화면에서는 오른쪽에 더 가깝게 */
-              width: 8px; /* 작은 화면에서는 스크롤바를 더 얇게 */
-            }
-          }
         `}</style>
         <Swiper
+          style={{ paddingBottom: '15rem' }} // 인라인 스타일로 marginBottom 추가
           ref={swiperRef}
           direction="vertical"
           modules={[Mousewheel, Scrollbar, EffectCoverflow]}
@@ -88,23 +90,26 @@ const Carousel: React.FC = () => {
           effect="coverflow"
           coverflowEffect={{
             rotate: 0,
-            stretch: spaceBetweenInPx, // 간격을 설정
-            depth: 200,
+            stretch: 0,
             modifier: 1,
             slideShadows: false,
           }}
-          slidesPerView={5}
+          slidesPerView={1.9}
           centeredSlides={true}
-          spaceBetween={spaceBetweenInPx} // 간격을 픽셀 값으로 설정
           scrollbar={{ draggable: true }}
-          simulateTouch={false} // 마우스 드래그 비활성화
-          speed={300} // 스크롤링 속도 조절 (기본값: 300ms)
-          className="h-full"
-          onSlideChange={(swiper) => updateSlideOpacityAndSize(swiper)}
-          onSwiper={(swiper) => updateSlideOpacityAndSize(swiper)}
+          simulateTouch={false}
+          speed={300}
+          className="h-full mb-20"
+          onSlideChange={updateSlideOpacityAndSize}
+          onSwiper={updateSlideOpacityAndSize}
         >
-          {newsData.map((news) => (
-            <SwiperSlide key={news.id} className="flex justify-center items-center">
+          {newsData.map((news, index) => (
+            <SwiperSlide
+              key={news.id}
+              className="flex justify-center items-center"
+              onClick={() => handleSlideClick(index)}
+              style={{ paddingTop: '14.7rem' }}
+            >
               <SavedNews />
             </SwiperSlide>
           ))}
